requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
websockets==12.0
httpx==0.28.1

# GitHub Copilot SDK
github-copilot-sdk==0.1.19

# File Processing
aiofiles==23.2.1

# PDF Generation
markdown==3.5.2
weasyprint==60.2
xhtml2pdf==0.2.16

# YAML Processing (Power Platform uses YAML for formulas)
pyyaml==6.0.1

# XML Processing (Power Platform uses XML for metadata)
lxml==5.4.0

# Utilities
python-dotenv==1.0.1
